- Tests + show pledge amount correctly for crowdfund
